Fix the patch command
