+++ requirements.txt
@@ -8,6 +8,7 @@ pytest-benchmark
 requests
 tabulate
 timm == 0.4.12
+# huggingface fx2trt support requires transformers == 4.12.1
 transformers == 4.12.1
 MonkeyType
 psutil
