+++ torchbenchmark/util/framework/huggingface/patches/0001-transformers-enable-fx.patch
@@ -2,7 +2,7 @@ diff --git a/file_utils.py b/file_utils.py
 index 8a19d298e..0d3ebf4fa 100644
 --- a/file_utils.py
 +++ b/file_utils.py
-@@ -412,7 +412,7 @@ def is_torch_tf32_available():
+@@ -306,7 +306,7 @@ def is_torch_tf32_available():
  _torch_fx_available = _torch_onnx_dict_inputs_support_available = False
  if _torch_available:
      torch_version = version.parse(importlib_metadata.version("torch"))
